refactor(cms): split writing queries
Move writing list and FTS search queries into cms/writing.ts so cms/index.ts keeps public exports and remaining page/project orchestration.

Verified with pnpm --filter @anipotts/lib typecheck, pnpm --filter @anipotts/lib test, pnpm --filter @anipotts/lib build, pnpm exec prettier --check packages/lib/src/cms/index.ts packages/lib/src/cms/writing.ts, git diff --check -- packages/lib/src/cms/index.ts packages/lib/src/cms/writing.ts, and pnpm dlx knip --no-exit-code.
