+++ packages/lib/src/cms/writing.ts
@@ -1,0 +1,100 @@
+import type { WritingSummary } from "@anipotts/types";
+import { and, desc, eq } from "drizzle-orm";
+import { getDB, getDrizzle, parseJsonArray } from "../db";
+import * as s from "../db/schema";
+import { logger } from "../logger";
+
+export async function fetchWriting(options?: {
+  published?: boolean;
+  limit?: number;
+}): Promise<WritingSummary[]> {
+  const db = getDrizzle();
+  if (db) {
+    try {
+      const conditions = [];
+      if (options?.published !== undefined) {
+        conditions.push(eq(s.thoughts.published, options.published));
+      }
+
+      const whereClause =
+        conditions.length > 0 ? and(...conditions) : undefined;
+
+      const results = await db
+        .select({
+          slug: s.thoughts.slug,
+          title: s.thoughts.title,
+          summary: s.thoughts.summary,
+          created_at: s.thoughts.created_at,
+          views: s.thoughts.views,
+          id: s.thoughts.id,
+          series_type: s.thoughts.series_type,
+          tags: s.thoughts.tags,
+        })
+        .from(s.thoughts)
+        .where(whereClause)
+        .orderBy(desc(s.thoughts.created_at))
+        .limit(options?.limit ?? 1000);
+      return results.map((row) => ({
+        slug: row.slug,
+        title: row.title,
+        summary: row.summary ?? "",
+        created_at: row.created_at ?? "",
+        views: row.views ?? undefined,
+        id: row.id ?? undefined,
+        series_type: row.series_type as WritingSummary["series_type"],
+        tags: parseJsonArray(row.tags),
+      }));
+    } catch (err) {
+      logger.warn("cms", "D1 fetchWriting() failed, using fallback", {
+        error: String(err),
+      });
+      return [];
+    }
+  }
+
+  return [];
+}
+
+export async function searchWriting(query: string): Promise<WritingSummary[]> {
+  const trimmed = query.trim();
+  if (!trimmed) return [];
+
+  const d1 = getDB();
+  if (d1) {
+    try {
+      // FTS5 phrase search keeps user input from being interpreted as MATCH syntax.
+      const phrase = `"${trimmed.replaceAll('"', '""')}"`;
+
+      // FTS5 search: must use raw SQL (Drizzle doesn't support FTS5 MATCH)
+      const { results } = await d1
+        .prepare(
+          `SELECT t.slug, t.title, t.summary, t.created_at, t.published_at, t.views, t.id, t.series_type, t.tags,
+                  rank
+           FROM thoughts_fts fts
+           JOIN thoughts t ON t.rowid = fts.rowid
+           WHERE thoughts_fts MATCH ?
+             AND (t.status = 'published' OR t.published = 1)
+           ORDER BY rank
+           LIMIT 20`,
+        )
+        .bind(phrase)
+        .all<Record<string, unknown>>();
+      return (results ?? []).map((row) => ({
+        slug: row.slug as string,
+        title: row.title as string,
+        summary: (row.summary as string) ?? "",
+        created_at: (row.created_at as string) ?? "",
+        published_at: row.published_at as string | undefined,
+        views: row.views as number | undefined,
+        id: row.id as string | undefined,
+        series_type: row.series_type as WritingSummary["series_type"],
+        tags: parseJsonArray(row.tags),
+      }));
+    } catch (err) {
+      logger.warn("cms", "D1 searchWriting() failed", { error: String(err) });
+      return [];
+    }
+  }
+
+  return [];
+}
